Revert "Table row styles should be in the same module as the theme."
This reverts commit 02e32164ef589164bd1da6f3956cf86a56dff842.
