+++ crates/core/src/theme/mod.rs
@@ -10,7 +10,6 @@ pub mod container;
 pub mod modal;
 pub mod picklist;
 pub mod scrollable;
-pub mod table_row;
 pub mod text;
 pub mod text_input;
 
@@ -21,7 +20,6 @@ pub use container::ContainerStyle;
 pub use modal::ModalStyle;
 pub use picklist::PickListStyle;
 pub use scrollable::ScrollableStyle;
-pub use table_row::TableRowStyle;
 pub use text_input::TextInputStyle;
 
 pub async fn load_user_themes() -> Vec<Theme> {
