+++ crates/core/src/widgets/style/table_row.rs
@@ -1,6 +1,5 @@
 //! Decorate content and apply alignment.
 use iced_core::{Background, Color};
-use crate::theme::Theme;
 
 /// The appearance of a table row.
 #[derive(Debug, Clone, Copy, Default)]
@@ -23,17 +22,26 @@ pub trait StyleSheet {
     fn hovered(&self) -> Appearance;
 }
 
-
-impl StyleSheet for Theme {
+impl StyleSheet for crate::theme::Theme {
     fn style(&self) -> Appearance {
-        Appearance::default()
+        Appearance {
+            text_color: Some(Color::WHITE),
+            background: None,
+            border_radius: 0.0,
+            border_width: 0.0,
+            border_color: Color::TRANSPARENT,
+            offset_right: 0.0,
+            offset_left: 0.0,
+        }
     }
 
     fn hovered(&self) -> Appearance {
-        Appearance::default()
+        Appearance {
+            background: None,
+            ..self.style()
+        }
     }
 }
-
 
 pub struct Default;
 impl StyleSheet for Default {
@@ -53,7 +61,6 @@ impl<'a> std::default::Default for Box<dyn StyleSheet + 'a> {
     }
 }
 
-
 impl<'a, T> From<T> for Box<dyn StyleSheet + 'a>
 where
     T: 'a + StyleSheet,
@@ -63,13 +70,120 @@ where
     }
 }
 
-
-/// A set of rules that dictate the style of a table row with assoicated style type.
-pub trait StyleSheetAssociated {
-    type Style: std::default::Default + Clone;
+/// A set of rules that dictate the style of a table row.
+pub trait StyleSheetWithStyle {
+    type Style: std::default::Default + Copy;
 
     fn appearance(&self, style: Self::Style) -> Appearance;
 
     /// Produces the style of a hovered table row.
     fn hovered(&self, style: Self::Style) -> Appearance;
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
+pub enum TableRowStyle {
+    #[default]
+    Default,
+    TableRowAlternate,
+    TableRowHighlife,
+    TableRowLowlife,
+    TableRowSelected,
+}
+
+// add default impl for Stylesheet
+impl StyleSheetWithStyle for crate::theme::Theme {
+    type Style = TableRowStyle;
+
+    fn appearance(&self, style: Self::Style) -> Appearance {
+        match style {
+            TableRowStyle::Default => Appearance {
+                text_color: Some(self.palette.normal.primary),
+                background: Some(Background::Color(self.palette.base.foreground)),
+                border_radius: 0.0,
+                border_width: 0.0,
+                border_color: Color::TRANSPARENT,
+                offset_left: 10.0,
+                offset_right: 25.0,
+            },
+            TableRowStyle::TableRowAlternate => Appearance {
+                background: Some(Background::Color(Color {
+                    a: 0.50,
+                    ..self.palette.base.foreground
+                })),
+                ..Appearance::default()
+            },
+            TableRowStyle::TableRowHighlife => Appearance {
+                text_color: Some(self.palette.normal.primary),
+                background: Some(Background::Color(Color {
+                    a: 0.30,
+                    ..self.palette.base.foreground
+                })),
+                border_radius: 0.0,
+                border_width: 0.0,
+                border_color: Color::TRANSPARENT,
+                offset_left: 0.0,
+                offset_right: 0.0,
+            },
+            TableRowStyle::TableRowLowlife => Appearance {
+                text_color: Some(self.palette.normal.primary),
+                background: Some(Background::Color(Color::TRANSPARENT)),
+                border_radius: 0.0,
+                border_width: 0.0,
+                border_color: Color::TRANSPARENT,
+                offset_left: 0.0,
+                offset_right: 0.0,
+            },
+            TableRowStyle::TableRowSelected => Appearance {
+                text_color: Some(self.palette.normal.primary),
+                background: Some(Background::Color(self.palette.normal.primary)),
+                border_radius: 0.0,
+                border_width: 0.0,
+                border_color: Color::TRANSPARENT,
+                offset_left: 0.0,
+                offset_right: 0.0,
+            },
+        }
+    }
+
+    fn hovered(&self, style: Self::Style) -> Appearance {
+        let appearance = self.appearance(style);
+
+        match style {
+            TableRowStyle::Default => Appearance {
+                background: Some(Background::Color(Color {
+                    a: 0.60,
+                    ..self.palette.normal.primary
+                })),
+                ..appearance
+            },
+            TableRowStyle::TableRowAlternate => Appearance {
+                background: Some(Background::Color(Color {
+                    a: 0.25,
+                    ..self.palette.normal.primary
+                })),
+                ..appearance
+            },
+            TableRowStyle::TableRowHighlife => Appearance {
+                background: Some(Background::Color(Color {
+                    a: 0.60,
+                    ..self.palette.normal.primary
+                })),
+                ..appearance
+            },
+            TableRowStyle::TableRowLowlife => Appearance {
+                background: Some(Background::Color(Color {
+                    a: 0.60,
+                    ..self.palette.normal.primary
+                })),
+                ..appearance
+            },
+            TableRowStyle::TableRowSelected => Appearance {
+                background: Some(Background::Color(Color {
+                    a: 0.60,
+                    ..self.palette.normal.primary
+                })),
+                ..appearance
+            },
+        }
+    }
 }
