+++ crates/core/src/widgets/widget/table_row.rs
@@ -1,5 +1,8 @@
 #![allow(clippy::type_complexity)]
-pub use super::super::style::table_row::{Appearance, StyleSheet, StyleSheetAssociated};
+
+use crate::widgets::style::table_row::TableRowStyle;
+
+pub use super::super::style::table_row::{Appearance, StyleSheet, StyleSheetWithStyle};
 use iced_native::{
     event, layout, mouse, overlay, renderer, widget, widget::Tree, Alignment, Clipboard, Element,
     Event, Layout, Length, Padding, Point, Rectangle, Shell, Widget,
@@ -20,7 +23,7 @@ where
     horizontal_alignment: Alignment,
     vertical_alignment: Alignment,
     style_sheet: Box<dyn StyleSheet + 'a>,
-    style: <Renderer::Theme as StyleSheetAssociated>::Style,
+    style: <Renderer::Theme as StyleSheetWithStyle>::Style,
     content: Element<'a, Message, Renderer>,
     on_press: Option<Box<dyn Fn(Event) -> Message + 'a>>,
 }
@@ -45,14 +48,16 @@ where
             horizontal_alignment: Alignment::Start,
             vertical_alignment: Alignment::Start,
             style_sheet: Default::default(),
-            style: <Renderer::Theme as StyleSheetAssociated>::Style::default(),
+            style: <Renderer::Theme as StyleSheetWithStyle>::Style::default(),
             content: content.into(),
             on_press: None,
         }
     }
-
-    /// Sets the style of the [`TableRow`].
-    pub fn style(mut self, style: <Renderer::Theme as StyleSheetAssociated>::Style) -> Self {
+    // pub fn style(mut self, style: <Renderer::Theme as StyleSheet>::Style) -> Self {
+    //     self.style_sheet = style.into();
+    //     self
+    // }
+    pub fn style(mut self, style: <Renderer::Theme as StyleSheetWithStyle>::Style) -> Self {
         self.style = style;
         self
     }
@@ -274,7 +279,10 @@ where
     }
 }
 
+//use grin_gui_core::theme::Theme as Custom;
+//pub trait Renderer: iced_native::Renderer<Theme = iced_native::Theme> {
 pub trait Renderer: iced_native::Renderer<Theme = crate::theme::Theme> {
+    //type Style: Default;
 
     #[allow(clippy::too_many_arguments)]
     fn draw<Message>(
@@ -283,6 +291,7 @@ pub trait Renderer: iced_native::Renderer<Theme = crate::theme::Theme> {
         layout: Layout<'_>,
         theme: &crate::theme::Theme,
         cursor_position: Point,
+        //style: TableRowStyle,
         style: &renderer::Style,
         style_sheet: &dyn StyleSheet,
         content: &Element<'_, Message, Self>,
