+++ src/gui/element/wallet/setup/wallet_list.rs
@@ -285,13 +285,13 @@ pub fn data_container<'a>(
 
         if selected_wallet {
             // selected wallet should be highlighted
-            table_row = table_row.style(grin_gui_core::theme::TableRowStyle::TableRowSelected);
+            table_row = table_row.style(grin_gui_core::widgets::style::table_row::TableRowStyle::TableRowSelected);
         } else {
             // contrast row styles to spice things up
             if pos % 2 == 0 {
-                table_row = table_row.style(grin_gui_core::theme::TableRowStyle::TableRowLowlife);
+                table_row = table_row.style(grin_gui_core::widgets::style::table_row::TableRowStyle::TableRowLowlife);
             } else {
-                table_row = table_row.style(grin_gui_core::theme::TableRowStyle::TableRowHighlife);
+                table_row = table_row.style(grin_gui_core::widgets::style::table_row::TableRowStyle::TableRowHighlife);
             }
         }
